sched: Fix loadavg calculations
This includes the alteration of the load calculation frequency
by Klaus Ripke <klaus@ripke.com> "LOAD_FREQ (4*HZ+61) avoids loadavg Moire"
http://ripke.com/loadavg/moire

But correctly alters the timings for 200 HZ devices as in the mobile
space, the original patch is meant for 100 HZ kernels on x86 and other archs.

Furthermore, the exponential constants are completely wrong for the same
reason, they were caluclated and hardcoded for 100 HZ.

From https://lkml.org/lkml/2013/2/1/93 , the exponential math is taken
to fully expand the macro behind those values and to be truly HZ independent.

The 64bit-ness of the above patch is avoided for our ARM purposes.