gpio_keys: force sync on KEY_POWER press
This should help in situations when the device or the OS is stalled but the kernel is alive and running.

Typically, user will press the power button to see if the device is responding
or to do a force restart.

By forcing a sync on KEY_POWER pressed, it will flush all the caches and prevent data corruption
and further more, prevent boot loop in these situations.

All the rest situations, when the device is NOT stalled and fully operational,
it will still not be a bad thing since earlysuspend.c is calling sys_sync() on suspend as well.