udp: fix behavior of wrong checksums
We have two problems in UDP stack related to bogus checksums :

1 ) We return -EGAIN to application even if receive queue is not empty.
    This breaks applications using edge trigger epoll()

2 ) Under UDP flood, we can loop forever without yielding to other
    processes, potentially hanging the host, especially on non SMP.

This patch is an attempt to make things better.

We might in the future add extra support for rt applications
wanting to better control time spent doing a recv() in a hostile
environment. For example we could validate checksums before queuing
packets in socket receive queue.

Change-Id: Ic25820228fd735e25a850951e059bf75ae39b653
Signed-off-by: Eruc Dumazet <edumazet@google.com>
Cc: Willem de Brujin <willemb@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>